move layout fragments
